+++ .env
@@ -1,0 +1,2 @@
+# Default environment (used if no ENVFILE and build mapping doesn't apply)
+BASE_URL=https://staging.supra.social
