+++ .env.production
@@ -1,0 +1,2 @@
+# Production environment
+BASE_URL=https://hub.supra.social
