+++ .env.staging
@@ -1,0 +1,2 @@
+# Staging environment
+BASE_URL=https://staging.supra.social
