+++ package.json
@@ -8,7 +8,11 @@
     "lint": "eslint .",
     "start": "react-native start",
     "test": "jest",
-    "build:android": "npx react-native bundle --platform android --dev false --entry-file index.js --bundle-output android/app/src/main/assets/index.android.bundle --assets-dest android/app/src/main/res && cd android && .\\gradlew assembleRelease && cd .."
+    "build:android": "npx react-native bundle --platform android --dev false --entry-file index.js --bundle-output android/app/src/main/assets/index.android.bundle --assets-dest android/app/src/main/res && cd android && .\\gradlew assembleRelease && cd ..",
+    "android:staging": "react-native run-android",
+    "android:prod:debug": "cross-env ENVFILE=.env.production react-native run-android",
+    "android:prod:release": "cross-env ENVFILE=.env.production react-native run-android --variant=release",
+    "android:prod:bundle": "cd android && cross-env ENVFILE=.env.production .\\gradlew bundleRelease && cd .."
   },
   "dependencies": {
     "@react-native-async-storage/async-storage": "^2.2.0",
@@ -19,6 +23,7 @@
     "react": "18.2.0",
     "react-native": "0.74.0",
     "react-native-app-auth": "^8.0.3",
+    "react-native-config": "^1.5.9",
     "react-native-safe-area-context": "^5.5.2",
     "react-native-screens": "^4.13.1",
     "tailwindcss": "^3.4.17",
@@ -39,7 +44,8 @@
     "jest": "^29.6.3",
     "prettier": "2.8.8",
     "react-test-renderer": "18.2.0",
-    "typescript": "5.0.4"
+    "typescript": "5.0.4",
+    "cross-env": "^7.0.3"
   },
   "engines": {
     "node": ">=18"
