+++ src/types/react-native-config.d.ts
@@ -1,0 +1,7 @@
+declare module 'react-native-config' {
+  export interface NativeConfig {
+    BASE_URL?: string;
+  }
+  const Config: NativeConfig;
+  export default Config;
+}
